+++ src/main/java/com/readum/domain/aiChat/exception/AiChatErrorCode.java
@@ -18,10 +18,7 @@ public enum AiChatErrorCode implements ErrorCode {
     AI_QUOTA_EXHAUSTED("AI 사용 한도가 소진되었습니다. 운영자에게 문의해 주세요."),
     AI_PROVIDER_ERROR("AI 응답 처리에 실패했습니다."),
     AI_PROVIDER_TRANSIENT("일시적인 AI 응답 오류입니다. 잠시 후 다시 시도해 주세요."),
-    AI_STREAM_INTERRUPTED("AI 응답이 중단되었습니다."),
-    SUMMARY_NOT_FOUND("아직 생성된 감상문이 없습니다."),
-    SUMMARY_IN_PROGRESS("감상문을 생성 중입니다. 잠시 후 다시 시도해 주세요."),
-    SUMMARY_GENERATION_FAILED("감상문 생성에 실패했습니다.");
+    AI_STREAM_INTERRUPTED("AI 응답이 중단되었습니다.");
 
     private final String message;
