+++ src/main/java/com/readum/domain/summary/service/SummarySearchService.java
@@ -1,10 +1,11 @@
-package com.readum.domain.aiChat.service;
+package com.readum.domain.summary.service;
 
-import com.readum.domain.aiChat.dto.SummaryResult;
 import com.readum.domain.aiChat.exception.AiChatErrorCode;
 import com.readum.domain.exception.ConflictException;
 import com.readum.domain.exception.NotFoundException;
 import com.readum.domain.exception.UnauthorizedException;
+import com.readum.domain.summary.dto.SummaryResult;
+import com.readum.domain.summary.exception.SummaryErrorCode;
 import com.readum.domain.user.exception.UserErrorCode;
 import com.readum.model.aiChat.repository.AiChatSessionRepository;
 import com.readum.model.summary.entity.Summary;
@@ -30,12 +31,12 @@ public class SummarySearchService {
                 .orElseThrow(() -> new NotFoundException(AiChatErrorCode.SESSION_NOT_FOUND));
 
         Summary summary = summaryRepository.findByAiChatSessionId(sessionId)
-                .orElseThrow(() -> new NotFoundException(AiChatErrorCode.SUMMARY_NOT_FOUND));
+                .orElseThrow(() -> new NotFoundException(SummaryErrorCode.SUMMARY_NOT_YET_CREATED));
 
         return switch (summary.getStatus()) {
             case COMPLETED -> SummaryResult.from(summary);
-            case IN_PROGRESS -> throw new ConflictException(AiChatErrorCode.SUMMARY_IN_PROGRESS);
-            case FAILED -> throw new ConflictException(AiChatErrorCode.SUMMARY_GENERATION_FAILED);
+            case IN_PROGRESS -> throw new ConflictException(SummaryErrorCode.SUMMARY_IN_PROGRESS);
+            case FAILED -> throw new ConflictException(SummaryErrorCode.SUMMARY_GENERATION_FAILED);
         };
     }
 }
