+++ src/test/java/com/readum/presentation/controller/aiChat/AiChatControllerTest.java
@@ -9,7 +9,6 @@ import com.readum.domain.aiChat.dto.MessageResult;
 import com.readum.domain.aiChat.dto.MessageStreamEvent;
 import com.readum.domain.aiChat.dto.SummaryDraftEligibility.IneligibleReason;
 import com.readum.domain.aiChat.dto.SummaryDraftEligibilityResult;
-import com.readum.domain.aiChat.dto.SummaryResult;
 import com.readum.domain.aiChat.exception.AiChatErrorCode;
 import com.readum.domain.aiChat.service.AiChatMessageSearchService;
 import com.readum.domain.aiChat.service.AiChatMessageSendService;
@@ -17,7 +16,9 @@ import com.readum.domain.aiChat.service.AiChatSessionCreateService;
 import com.readum.domain.aiChat.service.AiChatSessionSearchService;
 import com.readum.domain.aiChat.service.SummaryDraftSearchService;
 import com.readum.domain.aiChat.service.SummaryDraftService;
-import com.readum.domain.aiChat.service.SummarySearchService;
+import com.readum.domain.summary.dto.SummaryResult;
+import com.readum.domain.summary.exception.SummaryErrorCode;
+import com.readum.domain.summary.service.SummarySearchService;
 import com.readum.domain.exception.BadRequestException;
 import com.readum.domain.exception.ConflictException;
 import com.readum.domain.exception.NotFoundException;
@@ -466,7 +467,7 @@ class AiChatControllerTest {
     @Test
     void 감상문_조회_생성_요청_전이면_404() throws Exception {
         given(summarySearchService.findBySessionId(eq(1L), any()))
-                .willThrow(new NotFoundException(AiChatErrorCode.SUMMARY_NOT_FOUND));
+                .willThrow(new NotFoundException(SummaryErrorCode.SUMMARY_NOT_YET_CREATED));
 
         mockMvc.perform(get("/api/v1/ai-chat/sessions/1/summary")
                         .cookie(USER_SESSION_COOKIE))
@@ -477,7 +478,7 @@ class AiChatControllerTest {
     @Test
     void 감상문_조회_생성_중이면_409와_SUMMARY_IN_PROGRESS_메시지() throws Exception {
         given(summarySearchService.findBySessionId(eq(1L), any()))
-                .willThrow(new ConflictException(AiChatErrorCode.SUMMARY_IN_PROGRESS));
+                .willThrow(new ConflictException(SummaryErrorCode.SUMMARY_IN_PROGRESS));
 
         mockMvc.perform(get("/api/v1/ai-chat/sessions/1/summary")
                         .cookie(USER_SESSION_COOKIE))
@@ -488,7 +489,7 @@ class AiChatControllerTest {
     @Test
     void 감상문_조회_생성_실패면_409와_SUMMARY_GENERATION_FAILED_메시지() throws Exception {
         given(summarySearchService.findBySessionId(eq(1L), any()))
-                .willThrow(new ConflictException(AiChatErrorCode.SUMMARY_GENERATION_FAILED));
+                .willThrow(new ConflictException(SummaryErrorCode.SUMMARY_GENERATION_FAILED));
 
         mockMvc.perform(get("/api/v1/ai-chat/sessions/1/summary")
                         .cookie(USER_SESSION_COOKIE))
